tests/test_gRPC_service/service_tokenize/requirements.txt
fastmcp
grpcio
grpcio-tools
protobuf
llama-index
langchain
langchain-community
transformers
tokenizers
spacy
nltk
pydantic
python-dotenv
asyncio
uvloop
